+++ rs/new-dre/src/commands/heal.rs
@@ -11,9 +11,8 @@ impl ExecutableCommand for Heal {
     }
 
     async fn execute(&self, ctx: crate::ctx::DreContext) -> anyhow::Result<()> {
-        todo!("Implement once runner is migrated")
-        // let runner = ctx.runner();
-        // runner.network_heal(true).await
+        let runner = ctx.runner().await;
+        runner.network_heal().await
     }
 
     fn validate(&self, cmd: &mut clap::Command) {
