+++ rs/new-dre/src/ic_admin.rs
@@ -35,105 +35,6 @@ use tempfile::NamedTempFile;
 
 const MAX_SUMMARY_CHAR_COUNT: usize = 29000;
 
-#[derive(Clone, Serialize, PartialEq)]
-enum FirewallRuleModificationType {
-    Addition,
-    Update,
-    Removal,
-}
-
-impl Display for FirewallRuleModificationType {
-    fn fmt(&self, f: &mut fmt::Formatter) -> fmt::Result {
-        match self {
-            Self::Addition => write!(f, "add"),
-            Self::Update => write!(f, "update"),
-            Self::Removal => write!(f, "remove"),
-        }
-    }
-}
-
-#[derive(Clone, Serialize)]
-struct FirewallRuleModification {
-    change_type: FirewallRuleModificationType,
-    rule_being_modified: FirewallRule,
-    position: usize,
-}
-
-impl FirewallRuleModification {
-    fn addition(position: usize, rule: FirewallRule) -> Self {
-        Self {
-            change_type: FirewallRuleModificationType::Addition,
-            rule_being_modified: rule,
-            position,
-        }
-    }
-    fn update(position: usize, rule: FirewallRule) -> Self {
-        Self {
-            change_type: FirewallRuleModificationType::Update,
-            rule_being_modified: rule,
-            position,
-        }
-    }
-    fn removal(position: usize, rule: FirewallRule) -> Self {
-        Self {
-            change_type: FirewallRuleModificationType::Removal,
-            rule_being_modified: rule,
-            position,
-        }
-    }
-}
-
-struct FirewallRuleModifications {
-    raw: Vec<FirewallRuleModification>,
-}
-
-impl FirewallRuleModifications {
-    fn new() -> Self {
-        FirewallRuleModifications { raw: vec![] }
-    }
-
-    fn addition(&mut self, position: usize, rule: FirewallRule) {
-        self.raw.push(FirewallRuleModification::addition(position, rule))
-    }
-
-    fn update(&mut self, position: usize, rule: FirewallRule) {
-        self.raw.push(FirewallRuleModification::update(position, rule))
-    }
-
-    fn removal(&mut self, position: usize, rule: FirewallRule) {
-        self.raw.push(FirewallRuleModification::removal(position, rule))
-    }
-
-    fn reverse_sorted(&self) -> Vec<FirewallRuleModification> {
-        let mut sorted = self.raw.to_vec();
-        sorted.sort_by(|first, second| first.position.partial_cmp(&second.position).unwrap());
-        sorted.reverse();
-        sorted
-    }
-
-    fn reverse_sorted_and_batched(&self) -> Vec<(FirewallRuleModificationType, Vec<FirewallRuleModification>)> {
-        let mut batches: Vec<(FirewallRuleModificationType, Vec<FirewallRuleModification>)> = vec![];
-        let mut current_batch: Vec<FirewallRuleModification> = vec![];
-        let mut modtype: Option<FirewallRuleModificationType> = None;
-        for modif in self.reverse_sorted().iter() {
-            if modtype.is_none() {
-                modtype = Some(modif.clone().change_type);
-            }
-            if modtype.clone().unwrap() == modif.change_type {
-                current_batch.push(modif.clone())
-            } else {
-                batches.push((current_batch[0].clone().change_type, current_batch));
-                current_batch = vec![];
-                modtype = Some(modif.clone().change_type);
-            }
-        }
-        if !current_batch.is_empty() {
-            batches.push((current_batch[0].clone().change_type, current_batch))
-        }
-        batches
-    }
-}
-
 #[derive(Clone)]
 pub struct UpdateVersion {
     pub release_artifact: Artifact,
@@ -691,174 +592,6 @@ must be identical, and must match the SHA256 from the payload of the NNS proposa
 
         self.propose_run(command, options).await?;
         Ok(())
-    }
-
-    pub async fn update_firewall(
-        &self,
-        network: &Network,
-        propose_options: ProposeOptions,
-        firewall_rules_scope: &FirewallRulesScope,
-    ) -> Result<(), Error> {
-        let local_registry_path = local_registry_path(network);
-        let local_registry = LocalRegistry::new(local_registry_path, Duration::from_secs(10))
-            .map_err(|e| anyhow::anyhow!("Error in creating local registry instance: {:?}", e))?;
-
-        local_registry
-            .sync_with_nns()
-            .await
-            .map_err(|e| anyhow::anyhow!("Error when syncing with NNS: {:?}", e))?;
-
-        let value = local_registry
-            .get_value(&make_firewall_rules_record_key(firewall_rules_scope), local_registry.get_latest_version())
-            .map_err(|e| anyhow::anyhow!("Error fetching firewall rules for replica nodes: {:?}", e))?;
-
-        let rules = if let Some(value) = value {
-            FirewallRuleSet::decode(value.as_slice()).map_err(|e| anyhow::anyhow!("Failed to deserialize firewall ruleset: {:?}", e))?
-        } else {
-            FirewallRuleSet::default()
-        };
-
-        let rules: BTreeMap<usize, &FirewallRule> = rules.entries.iter().enumerate().sorted_by(|a, b| a.0.cmp(&b.0)).collect();
-
-        let mut builder = edit::Builder::new();
-        let with_suffix = builder.suffix(".json");
-        let pretty = serde_json::to_string_pretty(&rules).map_err(|e| anyhow::anyhow!("Error serializing ruleset to string: {:?}", e))?;
-        let edited: BTreeMap<usize, FirewallRule>;
-        loop {
-            info!("Spawning edit window...");
-            let edited_string = edit::edit_with_builder(pretty.clone(), with_suffix)?;
-            match serde_json::from_str(&edited_string) {
-                Ok(ruleset) => {
-                    edited = ruleset;
-                    break;
-                }
-                Err(e) => {
-                    warn!("Couldn't parse the input you provided, please retry. Error: {:?}", e);
-                }
-            }
-        }
-
-        let mut added_entries: BTreeMap<usize, &FirewallRule> = BTreeMap::new();
-        let mut updated_entries: BTreeMap<usize, &FirewallRule> = BTreeMap::new();
-        for (key, rule) in edited.iter() {
-            if let Some(old_rule) = rules.get(key) {
-                if rule != *old_rule {
-                    // Same key but different value meaning it was just updated
-                    updated_entries.insert(*key, rule);
-                }
-                continue;
-            }
-            // Doesn't exist in old ones meaning it was just added
-            added_entries.insert(*key, rule);
-        }
-
-        // Collect removed entries (keys from old set not present in new set)
-        let removed_entries: BTreeMap<usize, &FirewallRule> = rules.into_iter().filter(|(key, _)| !edited.contains_key(key)).collect();
-
-        let mut mods = FirewallRuleModifications::new();
-        for (pos, rule) in added_entries.into_iter() {
-            mods.addition(pos, rule.clone());
-        }
-        for (pos, rule) in updated_entries.into_iter() {
-            mods.update(pos, rule.clone());
-        }
-        for (pos, rule) in removed_entries.into_iter() {
-            mods.removal(pos, rule.clone());
-        }
-
-        let reverse_sorted = mods.reverse_sorted_and_batched();
-        if reverse_sorted.is_empty() {
-            info!("No modifications should be made");
-            return Ok(());
-        }
-        let diff = serde_json::to_string_pretty(&reverse_sorted).unwrap();
-        info!("Pretty printing diff:\n{}", diff);
-        /*if reverse_sorted.len() > 1 {
-            return Err(anyhow::anyhow!(
-                "Cannot currently apply more than 1 change at a time due to hash changes"
-            ));
-        }*/
-
-        //TODO: adapt to use set-firewall config so we can modify more than 1 rule at a time
-
-        async fn submit_proposal(
-            admin_wrapper: &IcAdminWrapper,
-            modifications: Vec<FirewallRuleModification>,
-            propose_options: ProposeOptions,
-            firewall_rules_scope: &FirewallRulesScope,
-        ) -> anyhow::Result<()> {
-            let positions = modifications.iter().map(|modif| modif.position).join(",");
-            let change_type = modifications[0].clone().change_type;
-
-            let mut file = NamedTempFile::new().map_err(|e| anyhow::anyhow!("Couldn't create temp file: {:?}", e))?;
-
-            let test_args = match change_type {
-                FirewallRuleModificationType::Removal => vec![
-                    "--test".to_string(),
-                    firewall_rules_scope.to_string(),
-                    positions.to_string(),
-                    "none".to_string(),
-                ],
-                _ => {
-                    let rules = modifications.iter().map(|modif| modif.clone().rule_being_modified).collect::<Vec<_>>();
-                    let serialized = serde_json::to_string(&rules).unwrap();
-                    file.write_all(serialized.as_bytes())
-                        .map_err(|e| anyhow::anyhow!("Couldn't write to tempfile: {:?}", e))?;
-                    vec![
-                        "--test".to_string(),
-                        firewall_rules_scope.to_string(),
-                        file.path().to_str().unwrap().to_string(),
-                        positions.to_string(),
-                        "none".to_string(),
-                    ]
-                }
-            };
-
-            let cmd = ProposeCommand::Raw {
-                command: format!("{}-firewall-rules", change_type),
-                args: test_args.clone(),
-            };
-
-            let output = admin_wrapper
-                .propose_run_inner(cmd, propose_options.clone(), true)
-                .await
-                .map_err(|e| anyhow::anyhow!("Couldn't execute test for {}-firewall-rules: {:?}", change_type, e))?;
-
-            let parsed: serde_json::Value = serde_json::from_str(&output)
-                .map_err(|e| anyhow::anyhow!("Error deserializing --test output while performing '{}': {:?}", change_type, e))?;
-            let hash = match parsed.get("hash") {
-                Some(serde_json::Value::String(hash)) => hash,
-                _ => {
-                    return Err(anyhow::anyhow!(
-                        "Couldn't find string value for key 'hash'. Whole dump:\n{}",
-                        serde_json::to_string_pretty(&parsed).unwrap()
-                    ))
-                }
-            };
-            info!("Computed hash for firewall rule at position '{}': {}", positions, hash);
-
-            let mut final_args = test_args.clone();
-            // Remove --test from head of args.
-            let _ = final_args.remove(0);
-            // Add the real hash to args.
-            let last = final_args.last_mut().unwrap();
-            *last = hash.to_string();
-
-            let cmd = ProposeCommand::Raw {
-                command: format!("{}-firewall-rules", change_type),
-                args: final_args,
-            };
-
-            admin_wrapper.propose_run(cmd, propose_options.clone()).await?;
-
-            Ok(())
-        }
-
-        // no more than one rule mod implemented currenty -- FIXME
-        match reverse_sorted.into_iter().last() {
-            Some((_, mods)) => submit_proposal(self, mods, propose_options.clone(), firewall_rules_scope).await,
-            None => Err(anyhow::anyhow!("Expected to have one item for firewall rule modification")),
-        }
     }
 }
